% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dina_vb.R
\name{dina}
\alias{dina}
\title{for the deterministic input noisy AND gate (DINA) model.}
\usage{
dina(
  X,
  Q,
  max_it = 500,
  epsilon = 1e-05,
  verbose = TRUE,
  delta_0 = rep(1, L),
  alpha_s = 1,
  beta_s = 1,
  alpha_g = 1,
  beta_g = 1
)
}
\arguments{
\item{X}{I by J binary matrix, item response data}

\item{Q}{J by K binary matrix, Q-matrix}

\item{max_it}{Maximum number of iterations (default: 500)}

\item{epsilon}{convergence tolerance for iterations (default: 10E-6)}

\item{verbose}{logical (default: TRUE)}

\item{delta_0}{L by 1 vector, initial value for variational parameter of Dirichlet
distribution.}

\item{alpha_s}{Positive scalar, variational parameter that determines the
shape of  prior beta distribution for slip parameter \emph{s_j}.}

\item{beta_s}{Positive scalar, variational parameter that determines the
shape of  prior beta distribution for slip parameter \emph{s_j}.}

\item{alpha_g}{Positive scalar, variational parameter that determines the
shape of  prior beta distribution for guessing parameter \emph{g_j}.}

\item{beta_g}{Positive scalar, variational parameter that determines the
shape of  prior beta distribution for guessing parameter \emph{g_j}.}
}
\value{
A list including:
\describe{
  \item{s_est}{the posterior mean of slip parameter.}
  \item{g_est}{the posterior mean of guessing parameter.}
  \item{s_sd}{the posterior standard diviation of slip parameter.}
  \item{g_sd}{the posterior standard diviation of guessing parameter.}
  \item{r_il}{the etimates of parameter for categorical distribution that is optimal variational posterior for \emph{\strong{z}_i}.}
  \item{alpha_s_ast}{the estimates of variational parameter for slip parameter}
  \item{beta_s_ast}{the estimates of variational parameter for slip parameter}
  \item{alpha_g_ast}{the estimates of variational parameter for guessing parameter}
  \item{beta_g_ast}{the estimates of variational parameter for guessing parameter}
  \item{pi_est}{the estimates of variational parameter for class-indicator \emph{z_i}}
  \item{delta_ast}{the estimates of hyper parameter for pi}
  \item{delta_sd}{the standard diviation of hyper parameter for pi}
  \item{l_lb}{the computed lower bound of each iteration}
  \item{att_pat_est}{the estimated attribute mastery patterns}
  \item{A}{all of the possible attribute mastery patterns}
  \item{Q}{the entered Q-matrix}
  \item{X}{the entered data matrix}
  \item{eta_lj}{the computed ideal responce}
  \item{m}{the number of performed iterations}
}
}
\description{
\code{dina()} returns variational Bayesian estimates for the DINA model.
}
\references{
Yamaguchi, K., & Okada, K. (2020). Variational Bayes inference
  for the DINA model. \emph{Journal of Educational and Behavioral
  Statistics}, 45(5), 569-597.
  \url{https://doi.org/10.3102/1076998620911934}.
}
