% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdcm_vb_diff_q.R
\name{hm_dcm}
\alias{hm_dcm}
\title{for the hidden Markov DCM.}
\usage{
hm_dcm(
  X,
  Q,
  A_0 = NULL,
  B_0 = NULL,
  delta_0 = NULL,
  ommega_0 = NULL,
  max_it = 500,
  epsilon = 1e-04,
  random_start = FALSE,
  model = "general",
  verbose = TRUE
)
}
\arguments{
\item{X}{T-length list whose each element contains I by J/T binary item response}

\item{Q}{T-length list whose each element contains J/T by K Q-matrix}

\item{A_0}{the value of hyperparameter A_0 (default: NULL)}

\item{B_0}{the value of hyperparameter B_0 (default: NULL)}

\item{delta_0}{the value of hyperparameter delta_0 (default: NULL)}

\item{ommega_0}{the value of hyperparameter ommega_0 (default: NULL)}

\item{max_it}{Maximum number of iterations (default: 500)}

\item{epsilon}{convergence tolerance for iterations (default: 1e-5)}

\item{random_start}{Logical, controls whether to print progress (default: "FALSE")}

\item{model}{"general" or "DINA" (default: "gengeral")}

\item{verbose}{logical (default: TRUE)}
}
\value{
A list including:
\describe{
  \item{theta_est}{the estimated conditional probability of taking a wrong response}
  \item{theta_sd}{the estimated standard deviation of theta}
  \item{pi_est}{the estimates of variational parameter for class indicator}
  \item{pi_sd}{the estimatd standard deviation of pi}
  \item{Tau_est}{the estimated transition probability}
  \item{Tau_sd}{the estimated standard deviation of the transition probability}
  \item{post_max_class}{the result of class analysis}
  \item{MAP_att_pat}{the MAP estimates of attribute mastery patterns}
  \item{att_master_prob}{the estimated attribute mastery probabilities}
  \item{EAP_att_pat}{the EAP estimates of attribute mastery patterns}
  \item{A_ast}{the computed A_ast}
  \item{B_ast}{the computed B_ast}
  \item{delta_ast}{the computed delta_ast}
  \item{ommega_ast}{the computed ommega_ast}
  \item{E_z_itl}{the resulted expectations of  attribute mastery pattern}
  \item{E_z_itl_z_itm1l}{the resulted expectations of  attribute mastery pattern at the time point t and t-1}
  \item{A_0}{the computed or entered value of A_0}
  \item{B_0}{the computed or entered value of B_0}
  \item{delta_0}{the computed or entered value of delta_0}
  \item{ommega_0}{the computed or entered value of ommega_0}
  \item{l_lb}{the computed lower bound of each iteration}
  \item{G_jt}{the computed G-matrix}
  \item{m}{the number of performed iterations}
}
}
\description{
\code{hm_dcm()} returns variational Bayesian estimates for the hidden
Markov DCM.
}
\references{
Yamaguchi, K., & Martinez, A. J. (2021). Variational Bayesian
  Inference Posterior Approximation Algorithm for Hidden Markov Diagnostic
  Classification Models. \url{https://doi.org/10.31234/osf.io/28jkf}.
}
