% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdina_vb.R
\name{mc_dina}
\alias{mc_dina}
\title{for the multiple-choice deterministic input noisy AND gate (MC-DINA) model.}
\usage{
mc_dina(X, Q, max_it = 500, epsilon = 1e-04, seed = 123, verbose = TRUE)
}
\arguments{
\item{X}{I by J binary matrix, item response data}

\item{Q}{J by (K+2) binary matrix, Q-matrix}

\item{max_it}{The maximum number of iterations (default: 500)}

\item{epsilon}{The convergence tolerance for iterations (default: 1e-5)}

\item{seed}{The seed value (default: 123)}

\item{verbose}{Logical, controls whether to print progress (default: TRUE)}
}
\value{
A list including:
\describe{
  \item{theta_est}{the estimate of the conditional response probability parameter \eqn{\Theta}}
  \item{theta_sd}{the posterior standard deviation of parameter \eqn{\Theta}}
  \item{pi_est}{the estimates of class mixing parameter \eqn{\pi}}
  \item{pi_sd}{the posterior standard deviations of class mixing parameter \eqn{\pi}}
  \item{a_ast}{the estimate of variational parameter \eqn{a^*}}
  \item{delta_ast}{the estimate of variational parameter \eqn{\delta^*}}
  \item{a_0}{the value of hyperparameter \eqn{a^0}}
  \item{delta_0}{the value of hyperparamter \eqn{\delta^0}}
  \item{l_lb}{the list of the values of evidence lower bound in each itertion}
  \item{att_pat_est}{the estimated attribute mastery patterns}
  \item{G-mat}{the computed G-matrix}
  \item{m}{the number of performed iterations}
  \item{seed}{the entered seed number}
}
}
\description{
\code{mc_dina()} returns variational Bayesian estimates for the MC-DINA model.
}
\references{
Yamaguchi, K. (2020). Variational Bayesian inference for the
  multiple-choice DINA model. \emph{Behaviormetrika}, 47(1), 159-187. \url{
  https://doi.org/10.1007/s41237-020-00104-w}
}
