% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_sim_Q}
\alias{mc_sim_Q}
\title{Artificial Q-matrix for MC-DINA model}
\format{
A matrix with components
\describe{
\item{column 1}{Item number}
\item{column 2}{Stem}
\item{column 3 to end}{attributes}
}
}
\usage{
mc_sim_Q
}
\description{
Artificial Q-matrix for a 30-item test measuring 5 attributes.
}
\references{
Yamaguchi, K. (2020). Variational Bayesian inference for the
  multiple-choice DINA model. \emph{Behaviormetrika}, 47(1), 159-187.
  \doi{10.1007/s41237-020-00104-w}
}
\keyword{datasets}
