% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdina_vb.R
\name{mcdina}
\alias{mcdina}
\title{for multiple-choice deterministic input noisy AND gate (MC-DINA) model.}
\usage{
mcdina(X, Q, max_it = 500, epsilon = 1e-05, seed = 123)
}
\arguments{
\item{X}{I by J binary matrix, item response data}

\item{Q}{J by K binary matrix, Q-matrix}

\item{max_it}{Maximum number of iterations}

\item{epsilon}{convergence tolerance for iterations}

\item{seed}{seed value}
}
\value{
A list including:
\describe{
  \item{theta_est}{the estimates of the conditional response probabilities}
  \item{theta_sd}{the standard deviations of theta}
  \item{pi_est}{the estimates of categorical distribution's parameters, which is the prior for the attribute mastery indicator}
  \item{pi_sd}{the standard deviations of pi}
  \item{r_il}{the expectation of the attribute mastery pattern conditioned by the categorical distribution}
  \item{a_ast}{the estimated variational Bayesian posterior for theta}
  \item{delta_ast}{the estimated variational Bayesian posterior for pi}
  \item{a_0}{the estimates of Dirichlet distribution's parameters, which is the prior for theta}
  \item{delta_0}{the estimates of Dirichlet distribution's parameters, which is the prior for pi}
  \item{l_lb}{the computed evidence of lower bound}
  \item{att_pat_est}{the estimated attribute mastery patterns}
  \item{A}{all of the possible attribute mastery patterns}
  \item{Q}{the entered Q-matrix}
  \item{X}{the entered data matrix}
  \item{G-mat}{the computed G-matrix}
  \item{m}{the number of performed iterations}
  \item{seed}{the entered seed number}
}
}
\description{
\code{mcdina()} returns variational Bayesian estimates for MC-DINA model.
}
\references{
Yamaguchi, K. (2020). Variational Bayesian inference for the
  multiple-choice DINA model. \emph{Behaviormetrika}, 47(1), 159-187. \url{
  https://doi.org/10.1007/s41237-020-00104-w}
}
