% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_Q_J30K3}
\alias{sim_Q_J30K3}
\title{Artificial Q-matrix for 30 items 3 attributes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 30 rows and 3 columns.
}
\source{
{artificially simulated}
}
\usage{
sim_Q_J30K3
}
\description{
this matrix represents an artificial Q-matrix for 30 items and 3 attributes
}
\keyword{datasets}
