% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_Q_J80K5.R
\docType{data}
\name{sim_Q_J80K5}
\alias{sim_Q_J80K5}
\title{Artificial Q-matrix (80 items)}
\format{
a 80 by 5 binary matrix
}
\usage{
sim_Q_J80K5
}
\description{
Artificial Q-matrix for a 80-item test measuring 5 attributes
}
\keyword{datasets}
