% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmdcm_vb_diff_q.R
\name{hmdcm_diff_q}
\alias{hmdcm_diff_q}
\title{estimates attributes mastery patterns for hidden Markov DCM.}
\usage{
hmdcm_diff_q(
  X,
  Q,
  A_0 = NULL,
  B_0 = NULL,
  delta_0 = NULL,
  ommega_0 = NULL,
  max_it = 500,
  epsilon = 0.001,
  Test_versions,
  test_order,
  model = "General",
  random_start = FALSE
)
}
\arguments{
\item{X}{N by J by T binary 3-dimension array, item response data}

\item{Q}{J by k by T binary 3-dimension array, Q-matrix}

\item{A_0}{the initial value of A}

\item{B_0}{the initial value of B}

\item{delta_0}{the initial value}

\item{ommega_0}{the initial value of}

\item{max_it}{Maximum number of iterations}

\item{epsilon}{convergence tolerance for iterations}

\item{Test_versions}{test version}

\item{test_order}{test order}

\item{model}{"general" or "DINA"}

\item{random_start}{logical}
}
\value{
A list including:
\describe{
  \item{theta_est}{}
  \item{theta_sd}{}
  \item{pi_est}{}
  \item{pi_sd}{}
  \item{Tau_est}{}
  \item{Tau_sd}{}
  \item{post_max_class}{}
  \item{MAP_att_pat}{}
  \item{att_master_prob}{}
  \item{EAP_att_pat}{}
  \item{A_ast}{}
  \item{delta_ast}{}
  \item{ommega_ast}{}
  \item{E_z_itl}{}
  \item{E_z_itl_z_itm1l}{}
  \item{A_0}{}
  \item{B_0}{}
  \item{delta_0}{}
  \item{ommega_0}{}
  \item{l_lb}{the computed lower bound of each iteration}
  \item{gamma_t_x_it}{}
  \item{log_zeta_sum}{}
  \item{A}{all of the possible attribute mastery patterns}
  \item{Q}{the entered Q-matrix}
  \item{X}{the entered data matrix}
  \item{G_jt}{the computed G-matrix}
  \item{m}{the number of performed iterations}
  \item{seed}{the entered seed number}
}
}
\description{
\code{hmdcm_diff_q()} returns variational Bayesian estimates for hidden
Markov DCM.
}
\references{
Yamaguchi, K., & Martinez, A. J. (2021). Variational Bayesian
  Inference Posterior Approximation Algorithm for Hidden Markov Diagnostic
  Classification Models. \url{https://doi.org/10.31234/osf.io/28jkf}.
}
