% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_Q_J80K5}
\alias{sim_Q_J80K5}
\title{Artificial Q-matrix for 80 items 5 attributes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 80 rows and 5 columns.
}
\source{
{artificially simulated}
}
\usage{
sim_Q_J80K5
}
\description{
Artificial Q-matrix for a 80-item test measuring 5 attributes
}
\keyword{datasets}
